autenticacion

+++ backend/router/router.py
@@ -1,4 +1,4 @@
-from fastapi import APIRouter, Response
+from fastapi import APIRouter, Response , Request
 from starlette.status import HTTP_201_CREATED, HTTP_204_NO_CONTENT, HTTP_401_UNAUTHORIZED
 from schema.user_schema import Userschema, DataUser
 from config.db import engine
@@ -8,6 +8,9 @@ from typing import List
 from schema.provincias_schema import ProvinciaSchema
 from model.provincias import provincias
 from model.users import users
+from middlewares.verify_token_route import verificarToken
+from functions_jwt import validate_token
+
 
 persona_prov = APIRouter()
 
@@ -17,7 +20,6 @@ def root():
     return {"message": "Im a fast api with a router"}
 
 #personas
-
 @persona_prov.get("/api/user", response_model=List[Userschema], tags=["User"]) 
 def get_users():
     with engine.connect() as conn:
@@ -25,13 +27,19 @@ def get_users():
         return result
 
 @persona_prov.get("/api/user/{user_id}", response_model=Userschema, tags=["User"])
-def get_user(user_id: str):
-        with engine.connect() as conn:
-            result= conn.execute(users.select().where(users.c.id == user_id)).first()
-            return result
+def get_user(user_id: str,request:Request):
+    a=verificarToken(request)
+    if(a != None):
+        return a
+    with engine.connect() as conn:
+        result= conn.execute(users.select().where(users.c.id == user_id)).first()
+        return result
 
-@persona_prov.post("/api/user", status_code=HTTP_201_CREATED, tags=["User"])
-def create_user(data_user: Userschema):
+@persona_prov.post("/api/user", status_code=HTTP_201_CREATED, tags=["User"] )
+def create_user(data_user: Userschema,request:Request):
+    a=verificarToken(request)
+    if(a != None):
+        return a
     with engine.connect() as conn:  #con with nos aseguramso que la conbecion de la base de dadatos se cierre
         new_user = data_user.dict()
         new_user["user_passw"] =generate_password_hash(data_user.user_passw, "pbkdf2:sha256:30", 30)
@@ -53,10 +61,13 @@ def user_login(data_user : DataUser):
 
 
 @persona_prov.put("/api/user/{user_id}", response_model=Userschema, tags=["User"])
-def update_user(data_update: Userschema, user_id:str):
+def update_user(data_update: Userschema, user_id:str,request:Request):
+    a=verificarToken(request)
+    if(a != None):
+        return a
     with engine.connect() as conn:
         encrypt_passw= generate_password_hash(data_update.user_passw, "pbkdf2:sha256:30",30)
-        conn.execute(users.update().values(name=data_update.name, username=data_update.username, user_passw= encrypt_passw).where(users.c.id== user_id))
+        conn.execute(users.update().values(name=data_update.name, username=data_update.username, user_passw= encrypt_passw,id_Provincia=data_update.id_Provincia).where(users.c.id== user_id))
         
         result = conn.execute(users.select().where(users.c.id == user_id)).first()
 
@@ -64,7 +75,10 @@ def update_user(data_update: Userschema, user_id:str):
     
 
 @persona_prov.delete("/api/user/{user_id}", status_code=HTTP_204_NO_CONTENT, tags=["User"])
-def delete_user(user_id: str):
+def delete_user(user_id: str,request:Request):
+    a=verificarToken(request)
+    if(a != None):
+        return a
     with engine.connect() as conn:
         conn.execute(users.delete().where(users.c.id == user_id))
         return  Response(status_code=HTTP_204_NO_CONTENT)
@@ -79,14 +93,20 @@ def get_provincias():
 
 
 @persona_prov.get("/api/provincias/{prov_id}", tags=["Provincias"])
-def get_single_province(prov_id: str):
+def get_single_province(prov_id: str,request:Request):
+        a=verificarToken(request)
+        if(a != None):
+            return a
         with engine.connect() as conn:
             result= conn.execute(provincias.select().where(provincias.c.id == prov_id)).first()
             return result 
         
         
 @persona_prov.post("/api/provincias", status_code=HTTP_201_CREATED, tags=["Provincias"])
-def create_province(data_province: ProvinciaSchema):
+def create_province(data_province: ProvinciaSchema,request:Request):
+    a=verificarToken(request)
+    if(a != None):
+            return a
     with engine.connect() as conn:  #con with nos aseguramso que la conbecion de la base de dadatos se cierre
         #new_province = data_province.dict()
         new_province= {"name": data_province.name}
@@ -97,7 +117,10 @@ def create_province(data_province: ProvinciaSchema):
 
 
 @persona_prov.put("/api/provincias/{prov_id}" , tags=["Provincias"])
-def update_provincia(data_update: ProvinciaSchema, prov_id: str):
+def update_provincia(data_update: ProvinciaSchema, prov_id: str,request:Request):
+    a=verificarToken(request)
+    if(a != None):
+        return a
     with engine.connect() as conn:
         conn.execute(provincias.update().values(name=data_update.name).where(provincias.c.id== prov_id))
         
@@ -107,7 +130,10 @@ def update_provincia(data_update: ProvinciaSchema, prov_id: str):
     
     
 @persona_prov.delete("/api/provincias/{prov_id}", status_code=HTTP_204_NO_CONTENT, tags=["Provincias"])
-def delete_user(prov_id: str):
+def delete_user(prov_id: str,request:Request):    
+    a=verificarToken(request)
+    if(a != None):
+        return a
     with engine.connect() as conn:
         conn.execute(users.delete().where(users.c.id_Provincia == prov_id))
         conn.execute(provincias.delete().where(provincias.c.id == prov_id))
